Fix template table and update content
The table wasn't rendering, so I updated that.

I also reorganized the content a bit, hoping that it's a bit easier to scan and understand. I also tried to be consistent with your main README. For example, the value to call the template uses an absolute path, which is what you have in the README

I also move the location-block to the end because it seemed different than the other metadata.

Use this as you will. And thanks for the converter!