Clarification of binary package type
If folks are less familiar with Linux or not paying close attention (like I was today :sweat_smile:), these download options will be confusing to them. The link under Debian is for Debian and derivatives, which includes Ubuntu and thus most desktop Linux distros. Linux links to an `.rpm`, which works for Fedora, CentOS, and RHEL. 

I'm offering a minor annotation to acknowledge that without too much explanation.